+++ ydb/core/tx/columnshard/blobs_action/tier/gc_actor.cpp
@@ -1,4 +1,5 @@
 #include "gc_actor.h"
+#include "object_key.h"
 
 #include <ydb/core/tx/columnshard/columnshard_private_events.h>
 
@@ -41,7 +42,7 @@ void TGarbageCollectionActor::Handle(NWrappers::NExternalStorage::TEvDeleteObjec
     TLogoBlobID logoBlobId;
     TString errorMessage;
     Y_ABORT_UNLESS(ev->Get()->Key);
-    AFL_VERIFY(TLogoBlobID::Parse(logoBlobId, *ev->Get()->Key, errorMessage))("error", errorMessage);
+    AFL_VERIFY(TObjectKey(GCTask->GetStorageId()).Parse(*ev->Get()->Key, logoBlobId, errorMessage))("error", errorMessage);
     BlobIdsToRemove.erase(logoBlobId);
     CheckFinished();
 }
@@ -64,9 +65,11 @@ void TGarbageCollectionActor::Bootstrap(const TActorContext& ctx) {
     for (auto&& i : GCTask->GetDraftBlobIds()) {
         BlobIdsToRemove.emplace(i.GetLogoBlobId());
     }
+
     for (auto&& i : BlobIdsToRemove) {
-        StartDeletingObject(i.ToString());
+        StartDeletingObject(TObjectKey(GCTask->GetStorageId()).Make(i));
     }
+
     TBase::Bootstrap(ctx);
     Become(&TGarbageCollectionActor::StateWork);
 }